+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Version 0.9 (2020-05-??)
 - enh, l4: chisquared detail plots to have region
 - enh, l4: chisquared + scatter of slopes improved aesthetics
 - enh, l4: chisquared added aggregates by world, continents
+- enh, l4: p3a chisquared+trends: html responsive
+- enh, l4: spin off p3a (simplified) and p3b (detailed) out of p3
 
 
 Version 0.8 (2020-05-13)